moved TypeScriptPlugin class into its own file.
This fixes issues with the mixed types of exports so other code (tests) can import TypeScriptPlugin.
